Applying a jack-hammer commit to pull in changes from AstroHuntsman/POCS
The two repos had diverged to a large extent and it was to much of a pain
to merge the two. This should be all the changes in Huntmsan that we are
going to pull into all at onces.

Note that this does not include some changes in the Hunstman version that
were done to `rs232.py` and `serial.py` so there might be a little work
after this (for the Huntsman folks).