Fix IntervalTimer so that it doesn't screw up the remainder; I found (#308)
Fix IntervalTimer so that it doesn't screw up the remainder; I found this happened when we paused for 30 seconds for a computer power-cycle. It is possible this is also why PAN006's telemetry board is not responding.

Extract SerialInputHandler out of telemetry_board.ino, adding the ability to specify the pin to be changed by name.

If power-cycling the computer, continue printing reports, which makes debugging easier.