Fix open_dome command so that it doesn't require that it be dark, (#333)
just that the weather be safe.

Re-write loop in AstrohavenDome._full_move so that it doesn't wait long
when reading from the dome. The old code was too slow, so the dome moved
in a little then stopped, then repeated that, but not fast enough for
closing to occur.

Fix a flakiness problem: once a limit has been reached (i.e. one shutter
all the way opened or all the way closed), the PLC will likely output
more than one of the corresponding limit characters before then
outputting the new stable character. So added code for reading the
state until it has reached a stable value.