Wait for tracking adjustment (#332)
The tracking adjustment for ioptron uses `move_ms_<dir>`, which places
the mount in `guide` mode (as indicated by the status) but is non-blocking.
If the tracking adjustment is too long (~ > 4-5 seconds) and no blocking
is done the machine will try to transition to next state. There is a
condition on the state machine that the mount must be in `tracking`
to `observe` so it gets stuck and then tries again. After 5 iterations
of same state without successful transition the machine will give up
and park the mount.

Here we just wait out for up to 30 seconds until back in `tracking`
mode.

Note: Ideally we should never have to have an adjustment of 4-5 seconds,
so not entirely sure what is going on there. Seems to happen most on
first adjustment after pointing but also depends on where mount is poining.
This is a problem with PAN001 that we have had and might not be an issue
on other units.

* Mount simulator will always report a fake offset
* Fixing call to `get_ms_offset` so it correctly uses axis (has no practical effect on ioptrons, which don't allow for differing rates between the axis)
* Renamed `test_run` -> `test_run_complete` so easier to run in isolation