AlreadyVisited Constraint
Only allow an object to be visited once. This is mostly useful for something like an engineering run where you have a set list of objects you want to observe once.

Note that the "best" object is still selected for each call, i.e. this is __not__ a sequential scheduler.

See also #402 for implementation.

Thanks @jermainegug !