Tool for sending observation information up to Datastore in bulk.
Note: this should be used for sending items that have not been stored.
The next PR will involve sending of observations as they are record so
we will have real-time information.

Note: This *does not* send the entity for the Unit as that should be part
of a registration script.