End of night shutdown improvements. (#407)
* End of night shutdown

A number of fixes and changes to account for situation when on valid
observations are found but we don't want to shut down the state machine.

* If scheduler has a list of `observations` then it `has_valid_observations`.
	which is True when starting.

Process:
* `scheduling` state calls `observatory.get_observation()`
* `get_observation` first checks if `has_valid_observations is False`
	if so it will reread the field list.
* `get_observation` either returns an observation or, if none found,
	calls `clear_available_observations`, making `has_valid_observations = False`.
* If no observation is returned to `scheduling` state, send to `parking` (then to `sleeping`).
* In `sleeping`, if `has_valid_observations is False` then sleep for an hour
	then try again (or shut down if `run_once` is specified).
  If `has_valid_observations is True` and it is safe and dark, shut down because
  	of unknown error.

Other changes:
* Remove `reread_fields_file` from dispatch scheduler as it is handled at
the `observatory` level by `observatory.get_observation`.
* Misc small fixes

* If no messaging then show `say` commands in log

* More shutdown cleanup

Note: This changes some of the state behavior for how parking and retrying
is handled. Housekeeping state will now only be called at the very end
of an observing run (either in the morning when unsafe or when too many
retry attempts). This moves some behavior from `sleeping` into `parked`.

* Fix the `should_retry` so it does not modify values
* Create a `reset_observing_run` method. Currenlty only resets number
of retries
* Other PR fixes

* Better name for private variable

* More updates

* Adding back the ability to reread fields file from the `get_observation`
call so that a user can do it manually rather than relying on an empty
observations list. But do it cleaner.
* Test the rereading of a fields with same entires only with a new higher
priority target
* Skip duplicate observation entries rather than stopping with assertion error.

* Removing a whole test as it seems to be a duplication of the one below
it for no reason and is also not testing what it says it should be testing.

* Clear fields_file when clearing observations
Explicitly clear observations for test of no targets

* Feedback from PR

* cleaning up code so parking logic is clearer