Add class ArduinoIO (#422)
pocs/sensors/arduino_io.py is based on peas/sensors.py, but
ArduinoIO handles only a single device, and writes readings
to a queue rather than to a database. File also contains
helpers for finding the arduinos, such as get_arduino_ports.

Moved protocol_arduinosimulator.py to a new folder,
pocs/serial_handlers. This will be the home of the other PySerial
handlers in the future.