Reduce flakiness of test_messaging.py
Use two multiprocessing.Event objects to coordinate the usage of the
messaging forwarder. In particular, once the forwarder has created
the sockets we can create the publisher and subscriber. And when
shutting down, we should wait until the sockets have been closed
else we may find that we can't create the ports for the next test.

Note that this reduces the flakiness on my laptop from perhaps
10% to around 0.5%, which is an improvement, but there is still
room to go.