Disable some tests when hardware is present (#433)
Disable some tests when hardware is present

Add option --without-hardware, complementing --with-hardware.
This allows the user to specify that tests are to be skipped
if they fail with hardware being present or if the user doesn't
want the hardware to be interacted with (e.g. when running
tests on a real unit).

Move parts of POCS/pocs/tests/conftest.py to POCS/conftest.py
in order to enable parsing of command line options without
having to specify pocs/tests on the commad line.

Mark 4 tests in test_observatory as disabled if hardware is present.

Ignore POCS/pytest.ini.