Give subscriber time to start
Sometimes receive_message() doesn't get the message
because the subscriber hasn't been properly hooked
up to the publisher through zeromq. This commit
combines the pub and sub fixtures into a single
pub_and_sub fixture that guarantees that the subscriber
is created first, and that there is a pause after
creating both the subscriber and the publisher.

Local testing finds this change to be 100% reliable
(i.e. ran the tests 1000 time each, no failures),
which is better than any previous attempt at fixing
issue #131.

Fixes #131.