Fix flakiness of test_messaging.py (#440)
Use two Event objects from a multiprocessing.SyncManager
to coordinate the usage of the messaging forwarder.
In particular, once the forwarder has created the sockets we
can create the publisher and subscriber. And when shutting
down, we should wait until the sockets have been closed else
we may find that we can't create the ports for the next test.

Sometimes receive_message() wasn't get the message
because the subscriber hadn't been properly hooked
up to the publisher through zeromq. This commit
combines the pub and sub fixtures into a single
pub_and_sub fixture that guarantees that the subscriber
is created first, and that there is a pause after
creating both the subscriber and the publisher.

Unfortunately, as the zeromq manual makes clear,
there isn't any guarantee about how long one must
wait before a subscriber is fully connected to a publisher.

Nonetheless, local testing finds this change to be 100%
reliable; specifically, I ran the tests 1000 time each with
no failures; that is better than any previous attempt at
fixing issue #131.

Fixes #131.