Fixing async calls for TheSkyX (#449)
Being a bad administrator and merging this so we can use it. We ran it last night for Huntsman run without problems. Comments from @jamessynge are good but will be applied in a future PR as per my comment above.

Note: This still needs some work to make all the calls consistent but
can be used for testing now.