Use threads instead of processes in test_pocs.py (#468)
Run pocs in a thread instead of in a separate process to make it easier
to share resources between the test thread and pocs. This was motivated
by PanMemoryDB, but I think it will help in general.

Add a Timeout class to pocs.utils and use in test_pocs to force the
ending of long running tests.

Shutdown pocs even if message not received; use a try-finally block
to ensure we send the shutdown message.

Use hardware module to get the list of simulator names, minus 'weather'.

Wait for pocs thread to stop even if an assertion has failed, to avoid
having two running at once (i.e. if the next test starts while the
previous test's thread is still running).