Add scripts for starting POCS automatically - WIP
These scripts are designed for running:

* ZeroMQ message forwarders
* PAWS Web Server
* PEAS Shell
* POCS Shell

They are each run in a separate terminal within
tmux, a terminal multiplexer.

These were originally developed for PAN008, and this PR
is for making them more generally applicable.

This is a WIP because I've not yet installed them on
PAN008 and tested them.