Improve logging and speed of startup scripts
Create a directory for each tmux_launch execution, used
by all of the startup scripts it launches.

Run the scripts for initializing each of the terminals
in a separate process and don't wait for them to complete.
This will speed the process of starting up the system.

Add -ex to all of the bash shebang lines so that the log
files show what was being executed, and so that bash
errors are detected.