Fix plot_weather command
plot_weather.py writes into /var/panoptes/plot_weather/, but
there is nothing that creates that directory, so this commit
adds that to plot_weather.py.

Running python directly from bash command line in a crontab
rule is failing, so this commit adds two shell scripts
for doing so: plot_weather.sh and update_indices.sh.

Updates README.md accordingly.

Updates the 'Current user:' lines in some scripts to use
the whoami command rather than $USER. It turns out that Posix
specifies $LOGNAME (i.e. login name) as the envvar to be set
to the name of the logged in user by the login command, not
$USER.