Adding a basic script for taking a bias frame.
* Adding a basic script for taking a bias frame.

This script takes the usb port (as reported by gphoto2) as the first
parameter and an optional second parameter for a filename. Default
filename is timestamped file prefixed with the word `bias-` that is
saved in current directory.

Changes camera settings to fastest possible shutter speed (1/4000s on
the EOS 100D), takes five exposures, then returns the camera to the
bulb setting.