Parking cleanups
* Cleanup to `parked` state to make it easier to understand
* Ctrl-c while pocs_shell is running will finish loop through state
machine rather than call park directly on the mount. This allows for
housekeeping (including uploading) of observations even when machine
is manually killed.

Note: in future might be worth adding an option such that when you
press Ctrl-c you are prompted for if you want immediate park and shutdown
or if should run through states.