Ensure that channel names don't have spaces.
Throws an exception if a channel name doesn't match
a regular expression, one which doesn't include spaces.
Fixes issue #205.

Greatly expand the class level documentation in PanMessaging.

Enforce that the value of a message is either a string
or a dict (that was already effectively required, but
now the failure message will be clear).