Add a config file for test_codestyle.py
pycodestyle supports reading config options from a file.
This change adds such a file and modifies test_codestyle.py
to reference it.

Removes ignoring of E402 (module level import not at top of file)
because we don't violate it now.

Sets the line length to 99, but continues to ignore E501 (line
too long) because we have a lot of those.