Fix some bugs and lots of warnings reported by pyflakes
Fixed some "unknown variable" errors, such as port when
it should have been self.port, or the corruption of
nickname by someone pasting the word position into the
middle of it.

Converted some string literals to raw string literals so
that the regular expression is not invalid; e.g 'PAN\d\d\d'
is invalid because \d is not a valid escape sequence. The
author likely meant \d to be the regular expression for
matching digits, which should either be '\\d' or r'\d'.

Removed unused imports. Marked some imports in __init__.py
files with "# pragma: no flakes" so that they can be exported
cleanly.

Replaced "from x import *" with explicit imports in libfli.py.

Removed many unused variables 'e' in statements of
the form: except Name as e.

Fixed uses of pytest.raises in test_social_messaging where
the asserts about the contents of the raised exception were
at the wrong indentation level, so were never executed.