If a dome simulator is explicitly added to the `pocs_local.yaml` file it breaks the testing, revealing deeper problems with the load config.
The issue is that the config is stored globablly, so with the move
to hardware that is created before the Observatory we now can have
the config loaded in a global way by hardware.

This is a temporary fix to an issue reported privately by @jamessynge.
A deeper examination of the config loading w.r.t. local files is warranted.