Remove support for the 'fileline' attribute in log formats (#678)
This non-standard attribute has been the cause of many problems,
so reverting back to using only the default attributes, in particular
filename and lineno.

Right align the filename in a 25 char field space (if it fits),
and pad the lineno to 3 chars with zero if lower than 100.
For most of our files, this will result in a fixed length prefix
before the message, thus making it somewhat easier to find the
message.

Adding ] as a marker/separator to log lines. The ] will be on all log lines,
between the (usually) fixed length prefix and the message.