Improve coverage of ArduinoIO and simulator
ArduinoSimulator now adds commands it receives
to the next output message that it produces.
Send command bytes to the simulator one at a time,
again to better simulate the reality that the
Arduino communication is serial... and to improve
code coverage.
Add "# pragma: no cover" to various unreachable
lines in the Arduino simulator code.

Test that ArduinoIO listens for messages on the
command topic and forwards the contained line
to the Arduino.

Use the new timeout_ms param of PanMessaging.receive_message
in ArduinoIO.handle_commands.

Add some comments, TODOs, and clean up some
code style issues.