Suppress the numpy DeprecationWarnings from astropy
A recent change to numpy has resulted in lots of warnings from
astropy. It appears that an upcoming change to astropy will fix
the code so that we don't get the DeprecationWarnings, but until
then this change tells pytest to not clutter the output with
all of those warnings.

FWIW, I'd prefer a fix that still showed one copy of the warning,
but wasn't sure how to achieve that with modest effort.