AC Power Check (#710)
The AC power reading will work similar to the weather in that will both check for a simulator (`power`) as well as an entry in the database.

Note that for now this assumes a database structure that does not exist, which is to say that we will need to modify how the sesnsors store their readings for this to work. This code will expect a `power.mains` top-level entry in the `current` collection.