Improve robustness of log formatter (#708)
* Improve robustness of log formatter

Attempt to undo the special casing that the standard
logging package applies to a dict as the second of
two args to a logging method (i.e. format_str, dict_value).

Annoyingly complicated. Makes me wonder if we should
simply require use of legacy formatting. Oh well, wasted
the effort to fix it, so let's try it out.

Add tests of log formatting code.

* Fix codestyle violation

* Remove dead code, fix coverage

* Respond to feedback, primarily adding docstrings.