Add a first docker 'helper' script
build-panuser-image.py creates an image that can be used for
running a command (defaults to bash) as PANUSER. This is useful
if you want to mount PANDIR (/var/panoptes) into the image.
By default docker images have 'root' as their USER, but that
means that the user is unlikely to match the owner of PANDIR.

Also adding scripts/install/create-panoptes-user.sh which does
what its name says. It can be used both in a docker build
and in the setup of a computer to run the PANOPTES scope.