Don't overshoot the closed position on the first shutter.
The feedback mechanism on the Astrohaven dome is rather limited:
it tells you whether a sensor on the lower shutter is tripped, but
nothing about the upper shutter. Since we were finding the dome
wasn't closing all the way reliably, overshooting seemed to
insure that the dome would close all the way. What we didn't
realize is that the variability in the straps and the sensor
could cause the overshoot to bend the lower shutter up and
over the pulley block on top of the fixed dome base, after
which we need to disassemble the dome to fix it. SIGH.

This was happening just for the first side to be closed (the
A side) as the other side couldn't close too far because it
would hit the top shutter of the first side. So, I've modified
the code so that it will stop immediately when it gets the
"closed" feedback from the first side, but then on the second
side will send one extra "close" command once the limit is
reached. Why? To avoid finding that the dome is almost closed,
but with a slight gap between the shutters.

Read SerialData params for a dome from a 'serial' dict.
As with AbstractSerialMount, this modifies AbstractSerialDome to
read the serial device configuration from an entry in the pocs*.yaml
file called 'serial' inside of the 'dome' entry.

Remove the default baudrate for the dome serial device from the
code and require that it be in the config.