Restoring the normal tolerance on the mpl image comparison.
Note that this is just to make things work. The image should be exactly
the same but my feeling is that the mpl backend or something is causing
a difference. The normal way to deal with this would be to upload the failed
test to some location where you could inspect the images and see why
different. Unfortunately the main way to do that is via the travis aretfacts,
which don't work for pull requests.

So, long story short, I'm not sure this is effectively testing this.