POCS Docker
This presents a number of changes for building and running POCS and friends
in a docker environment. As of this PR, starting the environment (via docker-compose)
will launch four separate services (container images):

* messaging-hub
* weather
* paws
* pocs

As of now, it will also map the following directories to the host:

* $PANDIR/json_store
* $PANDIR/logs
* $PANDIR/images
* $PANDIR/POCS

It also appropriately sets up the internal network that allows the services to communicate
with each other over zeromq and exposes port 8080 on the host machine for reaching
PAWS.

The environment can be started with a helper script:

```bash
$POCS/scripts/run_pocs_docker.sh [help|jupyterlab]
```

As of now there is one additonal option that can be passed (`jupyterlab`) which
will start a jupyterlab environment within the running pocs service, exposing port
8888 on the host for access.

The POCS repo is installed in development mode (i.e. `pip install -e .`) so that changes
to the `$POCS` dir on the host will be reflected in the running service.

To do still:
* Environment sensors a la the weather container.
* Convenience script to make it easier for non-technical users to install docker, authenticate,
pull images, etc.
* Lots more I'm sure.

Note: this runs all of the docker images in `privileged` mode, meaning they have access to the host.

Also:
* Dockerfile to create a `pocs-base` image that contains the POCS repository
installed on an ubuntu 18.04 base image. This includes an install of all the
dependencies, including astrometry.net. Some of the apt-packages have been
cleaned up or changed as part of this PR.
* docker-compose.yaml files for easily starting all of the containers and helpers.
* Bash script for starting environments.
* Helper scripts for starting services. The POCS script (`$POCS/scripts/docker/run_pocs.sh`) still needs some work.
* Small README updates (more likely to come).