Tracking correction threshold (#878)
The tracking corrections can be a minimum or maximum number of milliseconds.
For the minimum the tracking will be ignored, for the maximum the correction
will be set to the maximum.

Tracking correction thresholds are set in the config file.

Note: This has been running live on a number of units. Merging without review.