Dockerize
This will merge all the docker components into a new branch in the main repo. As part of milestone 0.7 these items will be cleaned up as a step towards making the docker version the default (i.e. `develop`) version for milestone 0.8.

This is being merged now because there are too many changes within one PR and it will be better to have new PRs that branch off this `docker` branch. This will also allow for installs on new units from the `panoptes` user rather than my personal branches.