Generate requirements.txt using pip-tools (#907)
Renames previous requirements.txt as requirements.in and generates
new requirements.txt file using pip-tools. This version of
requirements.txt serves as a flattened "lockfile" that is better
suited for reproducing deterministic builds. See example usage in
https://github.com/jazzband/pip-tools/#example-usage-for-pip-compile

To update, run:

    pip-compile requirements.in

---

Awesome, thanks @capsulecorplab! Looks like this could also solve my long-standing problem with maintaining the list in both `setup.py` (which we don't do in this branch) and `requirements.txt`.   I'll make an issue to get things converted over to `setup.py`, which is more in keeping with our other repositories.

Note that much of the work will be doing into what is currently the `docker` branch, so I will probably implement the `setup.py` changes there.

Thanks again for the PR!