reset_pocs update
Added two if statements and implemented the while loop. First if statement stops the user from getting an error when reset_pocs is run without a pocs instance or observatory class which causes "if self.pocs.observatory..." to fail.  Second if statement lets reset_pocs be run from the parked position.