Formatting, Mount Detection, Arduino Detection
Mount detection and crude Arduino detection. The Arduino detection uses a hardware difference between the control and camera boards. The control board has three ds18 temperature sensors, while the camera board does not. A sketcht hat tries to read these sensors is uploaded to either the camera board or control board. A disconnected sensor, and therefore all of the camera board's readings, will always have a temp of -127.00 C. This output allows the script to distinguish which Arduino is the camera board and which is the control board. This script is setup to prompt the user to automatically upload an Arduino sketch (power_board.ino / camera_board.ino) to the boards, however it's not fully automatic because of the way the arduino-cli is built. 

There are two limitations to the Arduino detection. 
1. The Arduino micro build will cause errors with the arduino-cli fqbn string system. This could be avoided with a simple if statement, but I need to see the output of either lsusb or arduino-cli board list on a system running a micro.
2. The arduino-cli sucks at uploading sketches in a simple way. Without setting up a config file, it requires a fqnb string, but even more annoyingly it needs to be called on a folder containing an Arduino sketch and non-Arduino files the cli creates during the compiling of a sketch. It won't make these files from a sketch created in the Arduino IDE, only when a sketch is made by the arduino-cli and compiled through the arduino-cli. I also think that these files are system-specific, but I am not sure. 

There is a simpler command just called "arduino" (I think it comes with the arduino IDE) that supports uploading sketches given only a sketch path and usb port. It would make things much simpler if it was included in the build.