Update install-pocs.sh (#987)
Thanks @jlibermann for the work on this!

I'm merging this as-is because I've run through the install a number of times on the Pi and other hardware.

---

* Update install-pocs.sh

* Test for ssh access to given github user and if fail, exit program with instructions on how to set up ssh.
* `panoptes-tutorials` always cloned.
* Check for clone failures.

* Update install-pocs.sh

* Move env vars to top of script.
* Cleanup old comments.

* Update install-pocs.sh

Slightly better help message.

* Update scripts/install/install-pocs.sh

Ah, right! Thanks!

Co-authored-by: Joshua Liberman <65671429+jlibermann@users.noreply.github.com>

* Update install-pocs.sh

* Don't use `exit` because we want to source script instead of run it.
* Switch to new docker group before pull.

* Update install-pocs.sh

Pull images as sudo otherwise it's annoying. :angry:

* Update install-pocs.sh

Better check for if line has been written to source file.

* Update install-pocs.sh

Double-quotes so we interpolate the rc sourcing.

* Update install-pocs.sh

Only clone repositories if installing as a developer

* Update install-pocs.sh

Fix comments in env file.

* Update install-pocs.sh

* Add cli option for unit install.
* Cleanup.
* Tidy the script.

Co-authored-by: Joshua Liberman <65671429+jlibermann@users.noreply.github.com>