** Unrelated to PR **
The `test_sdk_camera_not_found` is creating a camera class and then raising an exception with an invalid config, however the camera object is still in memory and not getting cleaned up by the garbage collector. This is causing a race condition on the `test_sdk_already_in_use` test, causing it to hang indefinitely in an odd way.  Simply moving the offending test below seems to fix the problem although it should probably be fixed more better.